% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgraph.R
\name{as.igraph}
\alias{as.igraph}
\alias{as.igraph.PRIMITIVE}
\alias{as.igraph}
\alias{as.igraph.sf}
\alias{as.igraph}
\alias{sc_as_igraph}
\title{Convert model to a graph.}
\usage{
\method{as.igraph}{PRIMITIVE}(x, ..., layout = TRUE)

\method{as.igraph}{sf}(x, ..., layout = TRUE)

sc_as_igraph(x, ..., layout = TRUE)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}

\item{layout}{keep the input vertex coordinates x-y as the graph layout, defaults to `TRUE` otherwise no layout is provided
## sf
## spatstat
## raw track data
## make up segments starting with the unjoin thing, that might show the way forward}
}
\description{
Only segments from an input model are extracted and used to build the graph.
}
\examples{
as.igraph(minimal_mesh)
}
