% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\name{as_tbl_graph}
\alias{as_tbl_graph}
\alias{as_tbl_graph.PRIMITIVE}
\alias{as_tbl_graph}
\alias{as_tbl_graph.sf}
\title{Convert to tidygraph.}
\usage{
\method{as_tbl_graph}{PRIMITIVE}(x, ...)

\method{as_tbl_graph}{sf}(x, ...)
}
\arguments{
\item{x}{model}

\item{...}{other args}
}
\description{
Convert to tidygraph.
}
\examples{
library(tidygraph)
as_tbl_graph(minimal_mesh)

## some kind of round trip

}
