% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\name{SC}
\alias{SC}
\alias{SC.tbl_graph}
\title{SC}
\usage{
\method{SC}{tbl_graph}(x, ...)
}
\arguments{
\item{x}{input model}

\item{...}{arguments passed to methods}
}
\description{
tbl_graph methods for the universal model
}
