% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgraph.r
\name{sc_as_igraph}
\alias{sc_as_igraph}
\title{Convert model to a graph.}
\usage{
sc_as_igraph(x, layout = TRUE, ...)
}
\arguments{
\item{x}{input object}

\item{layout}{keep the input vertex coordinates x-y as the graph layout, defaults to `TRUE` otherwise no layout is provided
## sf
## spatstat
## raw track data
## make up segments starting with the unjoin thing, that might show the way forward}

\item{...}{arguments passed to methods}
}
\description{
Only segments from an input model are extracted and used to build the graph.
}
